Implemented Product Creation
Implemented Product Creation and update crud page to show every product on database. Still have some things to change in the creation process, such as saving the image in the database, which will probably be made using aws. Next step is to implement edit and delete capabilities to the CRUD.