Purchase subtracts items from cart
Purchase now verifies the quantity of the product available in stock, and then subtracts item when purchase is effectuated. This is not only a main functionality of a e-commerce, but also makes the program more reliable and secure. Next step is to make a better order-history page and purchase page.