Updated order-history page
Update order-history page, showing some more information. The frontEnd is indeed awful, but it does work, and is not the priority of this project. Still, i'll probably make it better later. Don't know exactly the next step. Maybe i'll make a research bar or something like that, although i have no idea of how to do it. I'll study and figure the next step later.