Added edit-profile
Added edit-profile get and post mapping. Truth be told, there's not much to edit, since i intentionally made it so there isn't a lot of user data. Still, if there was, the system would still work, with only minor adjustments. Next step is pretty simple: add delete user functionality. Although it is simple, it's important to remember to delete the user's cart and maybe even the user's purchases (Although this might not be something a real e-commerce would do). Maybe, instead of deleting the account, we should just deactivate it to keep the data stored. Might research a bit on this.